Big package rename for consistency with other components
