+++ src/aiod_napari/evaluation/metrics.py
