+++ src/aiod_napari/inference/mask_export.py
@@ -18,8 +18,8 @@ from qtpy.QtWidgets import (
 )
 import skimage.io
 
-from ai_on_demand.utils import format_tooltip
-from ai_on_demand.widget_classes import SubWidget
+from aiod_napari.utils import format_tooltip
+from aiod_napari.widget_classes import SubWidget
 
 
 class ExportWidget(SubWidget):
