+++ src/aiod_napari/qcollapsible.py
