+++ src/aiod_napari/resources/CRICK_Brandmark_01_transparent.png
